home/ubuntu/project/backend/black-swan-sentinel/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background jobs and scheduling
apscheduler==3.10.4

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Data processing and analysis
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# Environment and configuration
python-dotenv==1.0.0

# CORS support
fastapi-cors==0.0.6

# Logging and monitoring
structlog==23.2.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

